refactor modal image build launch boundary
